Updated alot of pages and made styles jsx'x
