+++ src/App.js
@@ -1,7 +1,8 @@
 import { connect } from 'react-redux';
 import { useEffect } from 'react';
 import { BrowserRouter as Router, Switch, Route} from 'react-router-dom';
-import './App.css';import Login from './components/Login';
+import './App.css';
+import Login from './components/Login';
 import Home from './components/Home';
 import Navbar from './components/Navbar';
 import { getUserAuth } from './actions';
@@ -9,7 +10,7 @@ import { getUserAuth } from './actions';
 import { Resource } from "./components/Resource/Resource";
 import { Network } from "./components/Network/Network";
 import { Notifications } from "./components/notifications/notifications";
-import { Profile } from "./components/Profile/Profile";
+import Profile  from "./components/Profile.jsx";
 
 function App(props) {
   useEffect(() => {
