+++ src/components/Home.jsx
@@ -1,86 +1,35 @@
-import styled from "styled-components";
+import { Redirect } from 'react-router-dom';
+import { connect } from 'react-redux';
+
 import Leftside from './Leftside';
 import Main from './Main';
 import Rightside from './Rightside';
-import { Redirect } from 'react-router-dom';
-import { connect } from 'react-redux';
+
+
+import {
+  Container,
+  Section,
+  Layout
+} from "../styles/stylesHome";
  
 const Home = (props) => {
-    return (
-        <Container>
-            {!props.user && <Redirect to = '/' />}
-            <Section>
-                <h5>
-                    <a> Looking for resources? - </a>
-                    <p> Find resources in record time with Maine's Reentry Resources!</p>
-                </h5>
-            </Section>
-
-            <Layout>
-                <Leftside />
-                <Main />
-                <Rightside />
-            </Layout>
-        </Container>
+  return (
+    <Container>
+      {!props.user && <Redirect to = '/' />}
+      <Section>
+        <h5>
+          <a> Looking for resources? - </a>
+          <p> Find resources in record time with Maine's Reentry Resources!</p>
+        </h5>
+      </Section>
+        <Layout>
+          <Leftside />
+          <Main />
+          <Rightside />
+        </Layout>
+    </Container>
     );
 }
-
-const Container = styled.div`
-    padding-top: 52px;
-    max-width: 100%;
-`;
-
-const Content = styled.div`
-    max-width: 1120px;
-    margin-left: auto;
-    margin-right: auto;
-`;
-
-const Section = styled.section`
-    min-height: 50px;
-    padding: 16px 0;
-    box-sizing: content-box;
-    text-align: center;
-    text-decoration: underline;
-    display: flex;
-    justify-content: center;
-
-    h5 {
-        color: #0a66c2;
-        font-size: 14px;
-
-        a {
-            font-weight: 700;
-        }
-    }
-
-    p {
-        font-size: 14px;
-        color: #434649;
-        font-weight: 600;
-    }
-
-    @media (max-width: 768px) {
-        flex-direction: column;
-        padding: 0 5px;
-    }
-`;
-
-const Layout = styled.div`
-    display: grid;
-    grid-template-areas: "leftside main rightside";
-    grid-template-columns: minmax(0,5fr) minmax(0,12fr) minmax(300px,7fr);
-    column-gap: 25px;
-    row-gap: 25px;
-    margin: 25px 0;
-    /*grid-template-row: auto;*/
-
-    @media (max-width: 768px) {
-        display:  flex;
-        flex-direction: column;
-        padding: 0 5px;
-    }
-`;
 
 const mapStateToProps = (state) => {
     return {
