--- src/components/Profile/Popups.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-import "./pop.css";
-
- 
-export const Popups = props => {
-  return (
-    <div className="popup-box">
-      <div className="box">
-        <span className="close-icon" onClick={props.handleClose}>x</span>
-        {props.content}
-      </div>
-    </div>
-  );
-};
- 
-
