--- src/components/Profile/pop.css
@@ -1,34 +1,0 @@
-/* Popup style */
-.popup-box {
-    width: 100%;
-    top: 0;
-    left: 0;
-  }
-  .box::-webkit-scrollbar {
-    display: none;
-  }
-  .box {
-    position: relative;
-    z-index: 10;
-    width: 70%;
-    margin: 0 auto;
-    height: 250px;
-    background: #fff;
-    border-radius: 4px;
-    padding: 20px;
-    overflow: auto;
-  }
-  
-  .close-icon {
-    content: "x";
-    cursor: pointer;
-    position: absolute;
-    right: calc(17% - 100px);
-    top: calc(100vh - 100vh);
-    width: 25px;
-    height: 25px;
-    line-height: 20px;
-    text-align: center;
-    font-size: 20px;
-  }
-
