--- src/components/Profile/sidebar.css
@@ -1,216 +1,0 @@
-.sidebar{
-    display: flex;
-    flex-direction: column;
-    margin-top:4rem;
-    margin-left:20px;
-    width:400px;
-   
-}
-
-.sidebar__profile{
-  border-bottom: 1px solid rgb(212, 208, 208);
-  text-align: center;
-  background-color: #fff;
-  border-radius: .75rem;
-  padding-bottom:1rem;
-  overflow: hidden;
-}
-
-
-.sidebar__profile img{
-  width:100%;
-  height:50px;
-  margin-bottom: -30px;
-}
-.sidebar__profile .MuiAvatar-circular {
-  width: 50px;
-  height: 50px;
-  margin:0px auto;
-  object-fit: contain;
-
-}
-.profile__details{
-  display: flex;
-  justify-content: space-between;
-  margin-top: -10px;
-  padding-top:10px;
-
-}
-
-.profile__details > p{
-  margin: 5px 10px;
-  font-size: 13px;
-  color: rgb(100, 98, 98);
-  text-decoration: underline;
-  cursor: pointer;
-  font-size: 1rem;
-  font-weight: 500;
-}
-.profile__stats{
-  display: flex;
-  justify-content: space-between;
-  text-align: left;
-  font-size:12px;
-  padding: 0px 10px;
-  align-items: center;
-}
-.profile__stats > span {
-    font-size: 1rem;
-    font-weight: 500;
-    color: rgb(100, 98, 98);
-    text-decoration: underline;
-}
-
-.sidebar__recent{
-background-color: #fff;
-margin-top:1rem;
-border-radius: .5rem;
-padding-top:.5rem;
-
-}
-.sidebar__recent > h3{
-    font-size: 1.1rem;
-    font-weight: 500;
-    margin-left:10px;
-}
-
-.students_div{
-    display: flex;
- flex-direction: row;
- align-items: center;
-}
-
-
-.profile_details1 {
-    display: flex;
-    flex-direction: column;
-    align-items: left;
-    margin-left: 15px;
-}
-.profile_details1 > h4{
-    font-size: 1rem;
-    font-weight: 500;
-    margin-top:5px;
-    margin-bottom: -20px;
-}
-.profile_details1 > p{
-    font-size: .75rem;
-}
-
-.profiles_students .MuiAvatar-circular {
-    width: 50px;
-    height: 50px;
-    margin-left: 20px;
-    object-fit: contain;
-   
-  }
-  
-  .profile_details1 > button {
-    border-radius: 30px;
-    font-size: 1rem;
-    width:100px;
-    height: 35px;
-    color: rgb(59, 58, 58);
-    border: 2px solid #ccc;
-  }
-  .last_button {
-  
-  margin-top: 20px;   
-  text-align: center;
-  padding:10px;
-  width:100%;
-  border: 0;
-  outline: 0;
-  border: 1px solid rgb(219, 214, 214);
-  background-color: #fff;
-  font-size:1.1rem;
-  border-radius:  0 0 .5rem .5rem;
-  
-  }
-
-  .last_button:hover{
-    background-color: #f5f5f5;
-    cursor: pointer;
-  }
-
-
-  @media screen and (max-width: 600px) {
-    .sidebar{
-      width:100%;
-     
-    }
-    .sidebar__profile{
-      width:800px;
-    }
-    .sidebar__profile img{
-      width:100%;
-      height:50px;
-      margin-bottom: -30px;
-    }
-    .sidebar__profile .MuiAvatar-circular {
-      width: 50px;
-      height: 50px;
-      margin:0px auto;
-      object-fit: contain;
-  
-    }
-    .profile__details{
-      display: flex;
-      justify-content: space-between;
-      margin-top: -10px;
-      padding-top:10px;
-  
-  
-    }
-  
-    .profile__details > p{
-      margin: 5px 10px;
-      font-size: 13px;
-      color: rgb(100, 98, 98);
-      text-decoration: underline;
-      cursor: pointer;
-      font-size: 1rem;
-      font-weight: 500;
-    }
-    .profile__stats{
-      display: flex;
-      justify-content: space-between;
-      text-align: left;
-      font-size:12px;
-      padding: 0px 10px;
-      align-items: center;
-    }
-    .profile__stats > span {
-        font-size: 1rem;
-        font-weight: 500;
-        color: rgb(100, 98, 98);
-        text-decoration: underline;
-    }
-  
-    .sidebar__recent{
-      background-color: #fff;
-      margin-top:1rem;
-      border-radius: .5rem;
-      width:800px;
-  
-    }
-    .sidebar__recent > h3{
-        font-size: 1.1rem;
-        font-weight: 500;
-        margin-left:10px;
-    }
-  
-    .students_div{
-        display: flex;
-     flex-direction: row;
-     align-items: center;
-    }
-  
-  }
-
-  @media (max-width: 600px) {
-    .sidebar {
-      width: 135%;
-      margin-left:0px;
-    }
-  }
