--- src/components/Profile/userSlice.jsx
@@ -1,22 +1,0 @@
-// This is just for reference
-
-import { createSlice } from "@reduxjs/toolkit";
-
-export const userSlice = createSlice({
-    name: "user",
-    initialState: {
-        user: null,
-    },
-    reducers: {
-        loginUser: (state, action) => {
-            state.user = action.payload;
-        },
-        logoutUser: (state) => {
-            state.user = null;
-        },
-    },
-});
-export const { loginUser, logoutUser } = userSlice.actions;
-
-export const selectUser = (state) => state.user.user;
-export default userSlice.reducer;
