+++ src/styles/stylesLogin.jsx
@@ -1,0 +1,168 @@
+import styled from "styled-components";
+
+const Container = styled.div`
+  padding: 0px;
+
+`;
+
+const Nav = styled.nav`
+  max-width: 1128px;
+  margin: auto;
+  padding: 12px 0 16px 0;
+  display: flex;
+  align-items: center;
+  position: relative;
+  justify-content: space-between;
+  flex-wrap: nowrap;
+
+  & > a {
+    width: 135px;
+    height: 34px;
+
+    @media (max-width: 768px) {
+      padding: 0 5px;
+    }
+
+    img {
+      border-radius: 5px;
+      width:300px;
+
+      @media (max-width: 768px) {
+        border-radius: 2px;
+        width:200px;
+      }
+    }
+  }
+`;
+
+const Join = styled.button`
+  font-size: 16px;
+  padding: 10px 12px;
+  text-decoration: none;
+  color: rgba(0, 0, 0, 0.6);
+  margin-right: 12px;
+  font-weight: 600;
+  border-radius: 4px;
+
+  &:hover {
+    background-color: rgba(0, 0, 0, 0.08);
+    color: rgba(0, 0, 0, 0.9);
+    text-decoration: none;
+    cursor: pointer;
+  }
+`;
+
+const Guest = styled.button`
+  font-size: 16px;
+  padding: 10px 12px;
+  text-decoration: none;
+  color: rgba(0, 0, 0, 0.6);
+  margin-right: 12px;
+  font-weight: 600;
+  border-radius: 4px;
+
+  &:hover {
+    background-color: rgba(0, 0, 0, 0.08);
+    color: rgba(0, 0, 0, 0.9);
+    text-decoration: none;
+    cursor: pointer;
+  }
+`;
+
+const Section = styled.section`
+  display: flex;
+  align-content: start;
+  min-height: 700px;
+  padding-bottom: 138px;
+  padding-top: 40px;
+  padding: 60px 0;
+  position: relative;
+  flex-wrap: wrap;
+  width: 100%;
+  align-items: center;
+  margin: auto;
+
+  @media (max-width: 768px) {
+    margin: auto;
+    min-height: 0px;
+  }
+`;
+
+const Hero = styled.div`
+  width: 100%;
+  h1 {
+    padding-bottom: 0;
+    width: 55%;
+    font-size: 53px;
+    color: #2977c9;
+    font-weight: 200;
+    line-height: 70px;
+
+    @media (max-width: 768px) {
+      text-align: center;
+      font-size: 20px;
+      width: 100%;
+      line-height: 2;
+    }
+  }
+
+  img {
+    /*z-index: -1;*/
+    width: 500px;
+    border-radius: 50px;
+    position: absolute;
+    right: -1px;
+
+    @media (max-width: 768px) {
+      top: 230px;
+      display: block;
+      margin-left: auto;
+      margin-right: auto;
+      width: 50%;
+      border-radius: 15px;
+      position: initial;
+      height: initial;
+    }
+  }
+`;
+
+const Form = styled.div`
+  margin-top: 100px;
+  width: 408px;
+  @media (max-width: 768px) {
+    margin-top: 20px;
+  }
+`;
+
+const Google = styled.button`
+  display: flex;
+  justify-content: center;
+  background-color: #fff;
+  align-items: center;
+  height: 56px;
+  width: 100%;
+  border-radius: 28px;
+  box-shadow: inset 0 0 0 1px rgba(0 0 0 / 60%), inset 0 0 0 2px rgb(0 0 0 / 0%) inset 0 0 0 1px rgb(0 0 0 / 0%);
+  vertical-align: middle;
+  z-index: 0;
+  transition-duration: 167ms;
+  font-size: 20px;
+  color: rgba(0, 0 , 0, 0.6);
+
+  &:hover {
+    background-color: rgba(207, 207, 207, 0.25);
+    color: rgba(0, 0, 0, 0.75);
+    cursor: pointer;
+  }
+`;
+
+export {
+  Container,
+  Nav,
+  Join,
+  Guest,
+  Section,
+  Hero,
+  Form,
+  Google
+};
